requirements.txt
streamlit==1.28.0
google-api-python-client==2.104.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
pytz==2023.3
python-dotenv==1.0.0